Enhancements to MQTT
Fixed the topic name from settings is now being used. Was previously hardcoded. (Non breaking if you use the default topic, if not you would have to set it to the default value, gaggimate)
Cleanup of some parts of the code
Added a few more topics: pressure, puck-flow, flow and puck resistance.