Shot history account for post brewing drip (#357)
* Fix for shot recording to take into account weight stabilization after the shot is finished to allow all of the liquid to enter the cup before finalizing the shot data.
If the volumetric mode is active, the shot recording will now wait for the weight to stabilize before finalizing the shot data. This prevents premature finalization when the weight is still changing due to liquid settling. This is limited to a maximum of 3 seconds after the shot is finished or until the weight has stabilized.

* Changed to explicit safety checks instead of try/catch.

* Explicit checking

* Clean up

* Better handling of truncated data files (missing volume data specifically).

* Never take estimated volume

* Look for 12 columns (new) instead of 11.

* Revert utils handling for getting valid weight

---------

Co-authored-by: Jochen Niebuhr <kontakt@ju-hh.de>