Update HomekitBridgePlugin.cpp
Synchronization problem with the heating status caused by the use of atomic has been fixed.
Atomic is now only used for the switches, but not for the heating status, because Atomics was introduced to prevent race conditions when HomeKit (a separate process/task) attempts to control the GaggiMate controller.
However, the data flow is reversed for the heating status:
- Source: The controller triggers an event (boiler:heating:stable).
- Target: The HomeKit plugin receives the event and writes the value to the HomeSpan sensor (heatingSensor->setStability).
Since HomeSpan is programmed to be thread-safe internally and the setVal(value, true) method ensures that the value is copied safely to the HomeKit memory, no additional atomic is necessary here. There is no “decision loop” that could be interrupted.

In addition, homespan classes have been cleaned up again so that Apple Home accepts the devices without any problems.